@@ -2,10 +2,12 @@ import 'package:example/screens/prebuilt_widgets_demo_screen.dart';
 import 'package:flutter/material.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -13,7 +15,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: PrebuiltWidgetsScreen(),
+      home: const PrebuiltWidgetsScreen(),
     );
   }
 }
