@@ -2,9 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
 class LocationViewWidget extends StatelessWidget {
-  LocationViewWidget({
-    this.title,
-    this.color,
+  const LocationViewWidget({
+    super.key,
+    required this.title,
+    required this.color,
   });
 
   final Color color;
@@ -12,20 +13,18 @@ class LocationViewWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final DateTime _now = DateTime.now();
-    final String _currentDate = DateFormat('yyyy/MM/dd – kk:mm').format(_now);
-    return Container(
-      child: Column(
-        children: [
-          Text('${title.toUpperCase()}' ?? '',
-              style: TextStyle(
-                fontSize: 40,
-                fontWeight: FontWeight.w300,
-                color: color,
-              )),
-          Text(_currentDate.toString() ?? '')
-        ],
-      ),
+    final DateTime now = DateTime.now();
+    final String currentDate = DateFormat('yyyy/MM/dd – kk:mm').format(now);
+    return Column(
+      children: [
+        Text(title.toUpperCase(),
+            style: TextStyle(
+              fontSize: 40,
+              fontWeight: FontWeight.w300,
+              color: color,
+            )),
+        Text(currentDate.toString())
+      ],
     );
   }
 }
