@@ -6,14 +6,12 @@ part of 'details.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Details _$DetailsFromJson(Map json) {
-  return Details(
-    id: json['id'] as int,
-    weatherShortDescription: json['main'] as String,
-    weatherLongDescription: json['description'] as String,
-    icon: json['icon'] as String,
-  );
-}
+Details _$DetailsFromJson(Map json) => Details(
+      id: json['id'] as int,
+      weatherShortDescription: json['main'] as String,
+      weatherLongDescription: json['description'] as String,
+      icon: json['icon'] as String,
+    );
 
 Map<String, dynamic> _$DetailsToJson(Details instance) => <String, dynamic>{
       'id': instance.id,
