@@ -6,16 +6,14 @@ part of 'temperature.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Temperature _$TemperatureFromJson(Map json) {
-  return Temperature(
-    currentTemperature: (json['temp'] as num)?.toDouble(),
-    feelsLike: (json['feels_like'] as num)?.toDouble(),
-    tempMin: (json['temp_min'] as num)?.toDouble(),
-    tempMax: (json['temp_max'] as num)?.toDouble(),
-    pressure: json['pressure'] as int,
-    humidity: json['humidity'] as int,
-  );
-}
+Temperature _$TemperatureFromJson(Map json) => Temperature(
+      currentTemperature: (json['temp'] as num).toDouble(),
+      feelsLike: (json['feels_like'] as num).toDouble(),
+      tempMin: (json['temp_min'] as num).toDouble(),
+      tempMax: (json['temp_max'] as num).toDouble(),
+      pressure: json['pressure'] as int,
+      humidity: json['humidity'] as int,
+    );
 
 Map<String, dynamic> _$TemperatureToJson(Temperature instance) =>
     <String, dynamic>{
