@@ -6,13 +6,11 @@ part of 'wind.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Wind _$WindFromJson(Map json) {
-  return Wind(
-    speed: (json['speed'] as num)?.toDouble(),
-    deg: (json['deg'] as num)?.toDouble(),
-    gust: (json['gust'] as num)?.toDouble(),
-  );
-}
+Wind _$WindFromJson(Map json) => Wind(
+      speed: (json['speed'] as num).toDouble(),
+      deg: (json['deg'] as num).toDouble(),
+      gust: (json['gust'] as num?)?.toDouble(),
+    );
 
 Map<String, dynamic> _$WindToJson(Wind instance) => <String, dynamic>{
       'speed': instance.speed,
